+++ src/extension.ts
@@ -1,10 +1,12 @@
 import * as vscode from 'vscode';
 import { packageIdentifier } from './constants';
-import { runTests } from './utilities/test_runner';
+import { runTests, runAllTests } from './utilities/test_runner';
 import { runBuildRunner, runBuildRunnerDeleteConflicting, runClean, runPubGet } from './utilities/simple_commands_runner';
 
 export function activate(context: vscode.ExtensionContext) {
 	registerRunTestsContextMenu(context);
+
+	registerRunAllTestsContextMenu(context);
 
 	registerRunPubGetContextMenu(context);
 
@@ -17,7 +19,13 @@ export function activate(context: vscode.ExtensionContext) {
 
 function registerRunTestsContextMenu(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
-		vscode.commands.registerCommand(`${packageIdentifier}.runTests`, runTests)
+		vscode.commands.registerCommand(`${packageIdentifier}.runTests`, (uri) => runTests(uri))
+	);
+}
+
+function registerRunAllTestsContextMenu(context: vscode.ExtensionContext) {
+	context.subscriptions.push(
+		vscode.commands.registerCommand(`${packageIdentifier}.runAllTests`, runAllTests)
 	);
 }